requirements.txt
flask==2.0.1
python-dotenv==0.19.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
onnxruntime==1.8.1
sqlalchemy==1.4.23
flask-sqlalchemy==2.5.1
gunicorn==20.1.0 